+++ .gitignore
@@ -1,4 +1,6 @@
 HELP.md
+logs
+out
 target
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target
