--- src/main/java/com/skypro/shelterbot/enums/TrialPeriod.java
@@ -1,9 +1,0 @@
-package com.skypro.shelterbot.enums;
-
-public enum TrialPeriod {
-    CURRENT,
-    EXTENDED_14_DAYS,
-    EXTENDED_30_DAYS,
-    NOT_PASSED,
-    SUCCESS_PASSED
-}
