--- src/main/java/com/skypro/shelterbot/enums/WhichPet.java
@@ -1,6 +1,0 @@
-package com.skypro.shelterbot.enums;
-
-public enum WhichPet {
-    CAT,
-    DOG
-}
