--- src/main/java/com/skypro/shelterbot/exception/UserNotFoundException.java
@@ -1,5 +1,0 @@
-package com.skypro.shelterbot.exception;
-
-public class UserNotFoundException extends RuntimeException{
-
-}
