+++ src/main/java/com/skypro/shelterbot/model/Pet.java
@@ -1,14 +1,13 @@
 package com.skypro.shelterbot.model;
 
-import com.skypro.shelterbot.enums.WhichPet;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import javax.persistence.*;
-import java.util.List;
 
-@AllArgsConstructor
+import javax.persistence.*;
+
 @Data
+@AllArgsConstructor
 @NoArgsConstructor
 @Entity(name = "pets")
 public class Pet {
@@ -16,16 +15,17 @@ public class Pet {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private String name;
-    private int age;
 
     @Enumerated(value = EnumType.STRING)
-    private WhichPet whichPet;
+    private PetType type;
+    private String name;
+    private Integer age;
     private String breed;
 
-    @OneToMany(mappedBy = "pet")
-    private List<PhotoReport> photoReports;
-
-    public Pet(String name, int age, WhichPet whichPet, String breed) {
+    public Pet(PetType type, String name, Integer age, String breed) {
+        this.type = type;
+        this.name = name;
+        this.age = age;
+        this.breed = breed;
     }
 }
