+++ src/main/java/com/skypro/shelterbot/model/PetType.java
@@ -1,0 +1,15 @@
+package com.skypro.shelterbot.model;
+
+public enum PetType {
+    CAT("Кошка"),
+    DOG("Собака");
+    private final String description;
+
+    PetType(String description) {
+        this.description = description;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+}
