--- src/main/java/com/skypro/shelterbot/model/PhotoReport.java
@@ -1,25 +1,0 @@
-package com.skypro.shelterbot.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@AllArgsConstructor
-@Data
-@NoArgsConstructor
-@Entity
-public class PhotoReport {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String filePath;
-    private long fileSize;
-    private String mediaType;
-
-    @ManyToOne
-    @JoinColumn(name = "pet_id")
-    private Pet pet;
-}
