+++ src/main/java/com/skypro/shelterbot/model/Report.java
@@ -1,0 +1,27 @@
+package com.skypro.shelterbot.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity(name = "reports")
+public class Report {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "chat_id")
+    private User user;
+    private Long photoId;
+    private String text;
+    private LocalDateTime dateTime;
+    private Boolean isApproved;
+}
