+++ src/main/java/com/skypro/shelterbot/model/TrialPeriod.java
@@ -1,0 +1,9 @@
+package com.skypro.shelterbot.model;
+
+public enum TrialPeriod {
+    CURRENT,
+    EXTENDED_14_DAYS,
+    EXTENDED_30_DAYS,
+    NOT_PASSED,
+    SUCCESS_PASSED
+}
