+++ src/main/java/com/skypro/shelterbot/model/User.java
@@ -4,21 +4,28 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
+import javax.persistence.*;
 import java.sql.Timestamp;
+import java.util.List;
 
-@AllArgsConstructor
 @Data
+@AllArgsConstructor
 @NoArgsConstructor
 @Entity(name = "users")
 public class User {
 
     @Id
     private Long chatId;
-    private String shelter;
     private String name;
     private String phone;
-    private String mail;
+
+    @Column(name = "last_cmd")
+    private String lastCommand;
     private Timestamp registeredAt;
+
+    @OneToOne
+    private Pet pet;
+
+    @OneToMany(mappedBy = "user")
+    private List<Report> reports;
 }
