--- src/main/java/com/skypro/shelterbot/model/Volunteer.java
@@ -1,23 +1,0 @@
-package com.skypro.shelterbot.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@AllArgsConstructor
-@Data
-@NoArgsConstructor
-@Entity(name = "volunteers")
-public class Volunteer {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    private long chatId;
-    private String name;
-
-}
