--- src/main/java/com/skypro/shelterbot/repositories/PhotoReportRepository.java
@@ -1,11 +1,0 @@
-package com.skypro.shelterbot.repositories;
-
-import com.skypro.shelterbot.model.PhotoReport;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface PhotoReportRepository extends JpaRepository<PhotoReport,Long> {
-
-    Optional<PhotoReport> findPhotoReportById(Long petId);
-}
