--- src/main/java/com/skypro/shelterbot/repositories/ShelterCatUserRepository.java
@@ -1,11 +1,0 @@
-package com.skypro.shelterbot.repositories;
-
-import com.skypro.shelterbot.model.ShelterCatUser;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ShelterCatUserRepository extends JpaRepository<ShelterCatUser,Long> {
-
-    ShelterCatUser findShelterCatUserByChatId(Long chatId);
-
-    ShelterCatUser findShelterCatUserById(Long id);
-}
