--- src/main/java/com/skypro/shelterbot/repositories/ShelterDogUserRepository.java
@@ -1,11 +1,0 @@
-package com.skypro.shelterbot.repositories;
-
-import com.skypro.shelterbot.model.ShelterDogUser;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ShelterDogUserRepository extends JpaRepository<ShelterDogUser,Long> {
-
-    ShelterDogUser findShelterDogUserByChatId(Long chatId);
-
-    ShelterDogUser findShelterDogUserById(Long id);
-}
