--- src/main/java/com/skypro/shelterbot/repositories/VolunteerRepository.java
@@ -1,7 +1,0 @@
-package com.skypro.shelterbot.repositories;
-
-import com.skypro.shelterbot.model.Volunteer;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface VolunteerRepository extends JpaRepository<Volunteer,Long> {
-}
