+++ src/main/java/com/skypro/shelterbot/repository/ReportRepository.java
@@ -1,0 +1,13 @@
+package com.skypro.shelterbot.repository;
+
+import com.skypro.shelterbot.model.Report;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface ReportRepository extends JpaRepository<Report, Long> {
+    Optional<Report> findByUserChatId(Long chatId);
+    Optional<Report> findByPhotoId(Long photoId);
+}
