+++ src/main/java/com/skypro/shelterbot/repository/UserRepository.java
@@ -1,0 +1,15 @@
+package com.skypro.shelterbot.repository;
+
+import com.skypro.shelterbot.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByChatId(Long chatId);
+    Optional<User> findByPetId(Long petId);
+    void deleteByChatId(Long chatId);
+}
