+++ src/main/java/com/skypro/shelterbot/resource/StringConstants.java
@@ -1,0 +1,6 @@
+package com.skypro.shelterbot.resource;
+
+public interface StringConstants {
+
+    String HELP_TEXT = "Help text";
+}
