+++ src/test/java/com/skypro/shelterbot/ShelterBotApplicationTests.java
@@ -5,9 +5,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 class ShelterBotApplicationTests {
-
     @Test
     void contextLoads() {
     }
-
 }
